requirements.txt
langchain
langgraph
streamlit
faiss-cpu
sentence-transformers
PyPDF2
pdfplumber
pymupdf
pickle-mixin
